+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add the execution permission if possible
 * Show a specific message when user specify a script as game executable
 * Add lua scripting
+* Check for gdb presence
 
 ### Changed